나의 정보 / 링크 연결
Mypage
- 구분<student, professor>
- 로고 링크연결
- 나의 정보 클릭시 링크 잘못연결되어있던 것 해결
