+++ src/main/java/com/example/school_lms/Role/UserRole.java
@@ -1,5 +1,5 @@
 package com.example.school_lms.Role;
 
 public enum UserRole {
-    USER, ADMIN;
+    STUDENT, PROFESSOR;
 }
