+++ src/main/java/com/example/school_lms/controller/SessionLoginController.java
@@ -15,10 +15,6 @@ import org.springframework.validation.BindingResult;
 import org.springframework.validation.FieldError;
 import org.springframework.web.bind.annotation.*;
 
-import java.util.Enumeration;
-import java.util.HashMap;
-import java.util.Hashtable;
-import java.util.Map;
 
 @Controller
 @RequiredArgsConstructor
@@ -27,12 +23,14 @@ public class SessionLoginController {
 
     private final UserService userService;
 
+//    메인페이지
     @GetMapping(value = {"", "/"})
     public String mainUniversity(Model model, @SessionAttribute(name = "userId", required = false) Long userId) {
         model.addAttribute("loginType", "visang_university");
 
         User loginUser = userService.getLoginUserById(userId);
 
+        //로그인이 되어있으면 userdataName 뿌려주기
         if(loginUser != null) {
             model.addAttribute("userdataName", loginUser.getUserdataName());
         }
@@ -40,10 +38,10 @@ public class SessionLoginController {
         return "main";
     }
 
+//    회원가입 페이지
     @GetMapping("/join")
     public String joinPage(Model model) {
         model.addAttribute("loginType", "visang_university");
-//        model.addAttribute("pageName", "세션 로그인");
 
         model.addAttribute("joinRequest", new JoinRequest());
         return "join";
@@ -52,7 +50,6 @@ public class SessionLoginController {
     @PostMapping("/join")
     public String join(@Valid @ModelAttribute JoinRequest joinRequest, BindingResult bindingResult, Model model) {
         model.addAttribute("loginType", "visang_university");
-//        model.addAttribute("pageName", "세션 로그인");
 
         // 학번과 이름 중복 체크
         if(!userService.checkUserdataDuplicate(joinRequest.getUserdataNum(), joinRequest.getUserdataName())) {
@@ -72,10 +69,10 @@ public class SessionLoginController {
         return "redirect:/visang_university";
     }
 
+//    로그인 페이지
     @GetMapping("/login")
     public String loginPage(Model model) {
         model.addAttribute("loginType", "visang_university");
-//        model.addAttribute("pageName", "세션 로그인");
 
         model.addAttribute("loginRequest", new LoginRequest());
         return "login";
@@ -85,7 +82,6 @@ public class SessionLoginController {
     public String login(@ModelAttribute LoginRequest loginRequest, BindingResult bindingResult,
                         HttpServletRequest httpServletRequest, Model model) {
         model.addAttribute("loginType", "visang_university");
-//        model.addAttribute("pageName", "세션 로그인");
 
         User user = userService.login(loginRequest);
 
@@ -113,7 +109,6 @@ public class SessionLoginController {
     @GetMapping("/logout")
     public String logout(HttpServletRequest request, Model model) {
         model.addAttribute("loginType", "visang_university");
-//        model.addAttribute("pageName", "세션 로그인");
 
         HttpSession session = request.getSession(false);  // Session이 없으면 null return
         if(session != null) {
@@ -121,52 +116,4 @@ public class SessionLoginController {
         }
         return "redirect:/visang_university";
     }
-
-    @GetMapping("/info")
-    public String userInfo(@SessionAttribute(name = "userId", required = false) Long userId, Model model) {
-        model.addAttribute("loginType", "session-login");
-        model.addAttribute("pageName", "세션 로그인");
-
-        User loginUser = userService.getLoginUserById(userId);
-
-        if(loginUser == null) {
-            return "redirect:/visang_university/login";
-        }
-
-        model.addAttribute("user", loginUser);
-        return "info";
-    }
-
-    //교수 admin으로 잡음
-    @GetMapping("/admin")
-    public String adminPage(@SessionAttribute(name = "userId", required = false) Long userId, Model model) {
-        model.addAttribute("loginType", "session-login");
-        model.addAttribute("pageName", "세션 로그인");
-
-        User loginUser = userService.getLoginUserById(userId);
-
-        if(loginUser == null) {
-            return "redirect:/visang_university/login";
-        }
-
-        if(!loginUser.getRole().equals(UserRole.ADMIN)) {
-            return "redirect:/visang_university";
-        }
-
-        return "admin";
-    }
-//
-//    public static Hashtable sessionList = new Hashtable();
-//
-//    @GetMapping("/session-list")
-//    @ResponseBody
-//    public Map<String, String> sessionList() {
-//        Enumeration elements = sessionList.elements();
-//        Map<String, String> lists = new HashMap<>();
-//        while(elements.hasMoreElements()) {
-//            HttpSession session = (HttpSession)elements.nextElement();
-//            lists.put(session.getId(), String.valueOf(session.getAttribute("userId")));
-//        }
-//        return lists;
-//    }
 }
