+++ src/main/resources/application.properties
@@ -3,9 +3,10 @@ server.servlet.encoding.force-response=true
 
 #sql
 spring.datasource.driver-class-name=com.mysql.cj.jdbc.Driver
-spring.datasource.url=jdbc:mysql://yeil03.iptime.org:3306/team3?serverTimezone=Asia/Seoul
+#spring.datasource.url=jdbc:mysql://yeil03.iptime.org:3306/team3?serverTimezone=Asia/Seoul
+spring.datasource.url=jdbc:mysql://localhost:3306/testlms?serverTimezone=Asia/Seoul
 serverTimezone=UTC&characterEncoding=UTF-8
-spring.datasource.username=team3
+spring.datasource.username=root
 spring.datasource.password=1234
 spring.jpa.properties.hibernate.show_sql=true
 spring.jpa.properties.hibernate.format_sql=true
