+++ src/main/resources/templates/main.html
@@ -18,7 +18,7 @@
     <!-- 헤더 부분 -->
     <div class="header">
         <div class="logo">
-            <a href="main.html"><img src="/images/logo.png" alt="학교로고"></a>
+            <a th:href="@{/visang_university}"><img src="/images/logo.png" alt="학교로고"></a>
         </div>
         <div class="header_text">
             <div>
@@ -38,7 +38,7 @@
     <!-- 헤더 부분 -->
     <div class="header">
         <div class="logo">
-            <a href="main.html"><img src="/images/logo.png" alt="학교로고"></a>
+            <a th:href="@{/visang_university}"><img src="/images/logo.png" alt="학교로고"></a>
         </div>
         <div class="header_text">
             <div style="width: 150px;">
